% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pi.R
\name{calc_pi}
\alias{calc_pi}
\title{Monte carlo simulation to estimate pi.}
\usage{
calc_pi(n.iters = 1e+05)
}
\arguments{
\item{n.iters}{the number of Monte Carlo iterations.  Must exceed 1000.}
}
\value{
Monte Carlo based estimate of pi
}
\description{
Demonstration of example in 14.3.1 "What is pi".  This uses Monte Carlo simulation uses the ratio of randomly drawn points that are in a circle of radius r to a square with length r.
}
\examples{
# use a small number of iterations, likely get a poor estimate of pi
pi.est <- calc_pi(n.iters = 1000)
# get a more robust estimate, but very slow
pi.est <- calc_pi(n.iters = 10000000)



}
