% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.11.R
\name{fig2.11}
\alias{fig2.11}
\title{Dynamic behavior of the logistic model}
\usage{
fig2.11(r = 0.5, viewcode = F)
}
\arguments{
\item{r}{maximum rate of population growth}

\item{viewcode}{FALSE to run function (default), TRUE to run function and print code}
}
\description{
Generate figures 2.11 through 2.16 by specifying the maximum rate of population growth (r), and seeing accompanying cobweb diagram and the plot of population abundance vs. time
}
\examples{
# generate plot with "typical" dynamics
fig2.11(r = 0.5)
# generate plot with "complex" dynamics
fig2.11(r = 2.1)
#view plotting commands
print(viewcode = T)
}
