% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.1X.R
\name{fig2.1X}
\alias{fig2.1X}
\title{Generate figure 2.11 through 2.16: Dynamic behavior of logistic model}
\usage{
fig2.1X(r)
}
\description{
Generate figure 2.11 through 2.16: Dynamic behavior of logistic model
}
\examples{
#view plotting commands
print(fig2.1X)
# generate plot with "typical" dynamics
fig2.1X(r = 0.5)
# generate plot with "complex" dynamics
fig2.1X(r = 2.1)
}
