% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.2.R
\name{fig2.2}
\alias{fig2.2}
\title{Generate figure 2.2: Density Independent Model Comparison}
\usage{
fig2.2()
}
\description{
Generate figure 2.2: Density Independent Model Comparison
}
\examples{
#view plotting commands
print(fig2.2)
# generate plot
fig2.2()
}
