% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.3_2.4.R
\name{fig2.3_2.4}
\alias{fig2.3_2.4}
\title{Generate figures 2.3 and 2.4: Logistic Model}
\usage{
fig2.3_2.4()
}
\description{
Generate figures 2.3 and 2.4: Logistic Model
}
\examples{
#view plotting commands
print(fig2.3_2.4)
# generate plot
fig2.3_2.4()
}
