% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.6_7.R
\name{fig2.6_7}
\alias{fig2.6_7}
\title{Generate figures 2.6 and 2.7: Compare Logistic to Allee effect model}
\usage{
fig2.6_7()
}
\description{
Generate figures 2.6 and 2.7: Compare Logistic to Allee effect model
}
\examples{
#view plotting commands
print(fig2.6_7)
# generate plot
fig2.6_7()
}
