% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig3.1.R
\name{fig3.1}
\alias{fig3.1}
\title{Structured population Demo}
\usage{
fig3.1(viewcode = F)
}
\arguments{
\item{TRUE}{or FALSE (default) indicating whether to print the function code}
}
\value{
A list containing population growth rate and stable age distribution calculated via simulation and via eigenvalue / eigenvector analysis, elasticities, and a plot of transitent and stable model dynamics
}
\description{
Runs several examples of structured population analysis based on the 5-age-c.ass model illustrated in Chapter 3.  This includes the coding of structured models using both matrix and non-matrix notation, and use of eigenvalues and eigenvectors.  Returns the graphic in Figure 3.1.
}
\examples{
# Run the code, and print output at console
fig3.1()

# Run the code, assign output to an object
output <- fig3.1()
print(output)

# View the code
fig3.1(viewcode = T)

}
