% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig4.1.R
\name{fig4.1}
\alias{fig4.1}
\title{Generate figure 4.1: What is model stability?}
\usage{
fig4.1(stability = "stable")
}
\arguments{
\item{stability}{character string that is either "stable" or "unstable"}
}
\value{
a list object containing jacobian matrix and dominant eigenvalue
}
\description{
This function uses a continuous-time lotka-volterra competition model to introduce the concept of stability, how to use the Adams-Bashford method to simulate a model with differential equations, and the calculation of Jacobian matrices and dominant eigenvalue.  The user can ask for a "stable" model configuration or an "unstable" model configuration
}
\examples{
#view  commands
print(fig4.2)
# generate plot with a stable model
fig2.2("stable")
# generate a plot with an unstable model
fig2.2("unstable")
}
