% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig5.2.R
\name{fig5.2}
\alias{fig5.2}
\title{Demographic stochasticity}
\usage{
fig5.2()
}
\description{
Generates figure 5.2, by simulating the time dynamics of a population demographic stochasticity (where the number of births and deaths each year is random because the population size is small) is important. Each time the simulation is run, a different outcome will be seen.
}
\examples{
# View the code
print(fig5.2)
# See a randomized version of Figure 5.2
fig5.2()
}
