% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig5.8.R
\name{fig5.8}
\alias{fig5.8}
\title{Stochastic Allee Effect Models}
\usage{
fig5.8()
}
\description{
Compare scenarios shown in Figure 5.8 where (1) stochasticity caused a population to drop below the Allee threshold and therefore collapse and (2) where stochasticity saved a population by allowing it to increase above the extinction threshold.
}
\examples{
View code
print(fig5.8)
Re-create a version of Figure 5.8
fig5.8()
}
