% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig8.3.R
\name{fig8.3}
\alias{fig8.3}
\title{Poison Maximum Likelihood Estimation, single observation}
\usage{
fig8.3(viewcode = F)
}
\arguments{
\item{viewcode}{TRUE or FALSE (default) indicating whether to print the function code}
}
\value{
a list object containing the maximum likelihood estimate of r and the confidence interval
}
\description{
This function generates  Figure 8.3 and returns the maximum likelihood estimate of r and the confidence interval
}
\examples{
# generate plot shown in Fig 8.3
fig8.3()
#view  commands
fig8.3(viewcode = F)
}
