% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig8.4.R
\name{fig8.4}
\alias{fig8.4}
\title{Poison Maximum Likelihood Estimation, multiple observations}
\usage{
fig8.4(viewcode = FALSE)
}
\arguments{
\item{viewcode}{TRUE or FALSE (default) indicating whether to print the function code}
}
\value{
a list object containing the maximum likelihood estimate of r and the confidence interval
}
\description{
This function generates  Figure 8.4 and returns the maximum likelihood estimate of r and the confidence interval
}
\examples{
# generate plot shown in Fig 8.4, and return fit and confidence interval
fig8.4()
# View code
fig8.4(viewcode = TRUE)
}
