% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig8.8.R
\name{fig8.8}
\alias{fig8.8}
\title{Fit Survival Function to Goby Data}
\usage{
fig8.8(viewcode = FALSE)
}
\arguments{
\item{viewcode}{TRUE or FALSE (default) indicating whether to print the function code}
}
\value{
a list object containing the maximum likelihood estimate of pmax and alpha
}
\description{
This function generates Figure 8.8 by estimating the maximum likelihood parameter values of the survival function in Eq. 8.10
}
\examples{
# generate plot in Fig 8.8 and get maximum likelihood estimated
fit <- fig8.8()
print(fit)
#view  commands
fig8.8(viewcode = TRUE)
}
