% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gobies.R
\docType{data}
\name{gobies}
\alias{gobies}
\title{Post settlement goby survivorship}
\format{
A data frame with 15 observations, each representing an experimental trial
\describe{
  \item{initial_number}{Number of gobies initially placed in each experimental unit.}
  \item{final_number}{Number of gobies present at the end of the experiment.}
  }
}
\source{
Forrester, G.E. 1995  Strong Density-Dependent Survival and Recruitment
Regulate the Abundance of a Coral Reef Fish. Oecologia 103(3): 275 - 282.
}
\usage{
gobies
}
\description{
Experimental results from Forrester (1995) documenting effect of density on survivorship of Bridled goby (Coryphopterus glaucofraenum)
on artificial reefs.
}
\keyword{datasets}
