% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-housewren.R
\docType{data}
\name{housewren}
\alias{housewren}
\title{Population Dynamics of House Wrens}
\format{
A data frame with 44 observations, each representing a year
\describe{
  \item{Year}{Calendar year.}
  \item{Count}{Number of male territories or breeding pairs observed per year.}
  }
}
\source{
Kendeigh, S. C.  1982. Bird populations in Eastern Illinois: Fluctuations, variations, and development over a half-century. Illinois Biological Monographs. Vol. 52. 136 pp.
}
\usage{
housewren
}
\description{
Annual estimated abundance of breeding House Wren (Troglodytes aedon), in Eastern Illinios at William Trelease Woods, a 24 hectare forested plot owned by the University of Illinois.
}
\keyword{datasets}
