% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/killer_whale_extinction.R
\name{killer_whale_extinction}
\alias{killer_whale_extinction}
\title{Extinction risk of Southern Resident Killer Whales}
\usage{
killer_whale_extinction(model.type = "base")
}
\arguments{
\item{model.type}{is character that is either 'base', 'autocorrelation', or 'uncertainty', associated with the three scenarios in Table 5.1.}
}
\value{
the estimated probability of extinction over the next 100 years, and plots Figures 5.5 and 5.6.
}
\description{
Perform simulations to evaluate the risk that southern resident killer whales will drop below the quasi-extinction threshold (4) over the next 100 years, given the data on total killer whale counts.  The model simulates 10,000 random population runs, under three scenarios
\itemize{
\item "base", where parameters are assumed to be known and the lambdas are independent
\item "autocorrelation" where the log-lambdas are presumed to be autocorrelated
\item "uncertainty" where we admit uncertainty in the estimates of the mean and standard deviation of the log lambdas
}
}
\examples{
# View the code
print(killer_whale_extinction)
# Run the base model
extinct.base <- killer_whale_extinction(model.type = "base")
# create the uncertainty model
extinct.uncertainty <- killer_whale_extinction(model.type = "uncertainty")
}
