% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-killerwhales.R
\docType{data}
\name{killerwhales}
\alias{killerwhales}
\title{Population Dynamics of Killer Whales}
\format{
A data frame with 45 observations, each representing a year
\describe{
  \item{Year}{Calendar year.}
  \item{Count}{Total summed abundance of J, K and L pods.}
  }
}
\source{
Center for Whale Research https://www.whaleresearch.com/
}
\usage{
killerwhales
}
\description{
Population abundance of Southern Resident Killer Whales, summed for J, K, and L pods.  Data are courtesy of the
Center for Whale Research: https://www.whaleresearch.com/
}
\keyword{datasets}
