% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotka_volterra.R
\name{lotka_volterra}
\alias{lotka_volterra}
\title{Simulate Lotka-Volterra Predator-Prey Model}
\usage{
lotka_volterra(perturb = 1.05)
}
\arguments{
\item{perturb}{is a multiplier that is used to perturb the model away from equilibrium.  The default 1.05 perturbs by 5 percent away from equilibrium}
}
\value{
a list containing the jacobian matrix and dominant eigenvalue, and creates figure 4.8
}
\description{
Simulate the time dynamics of the Lotka-Volterra model, and recreate figure 4.8
}
\examples{
# View the code
print(lotka_volterra)
# Recreate figure 4.8 exactly
lotka_volterra(perturb = 1.05)
# create a larger initial perturbation away from equilibrim
lotka_volterra(perturb = 1.25)
}
