% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotkaVolterra.R
\name{lotkaVolterra}
\alias{lotkaVolterra}
\title{Simulate Lotka-Volterra Predator-Prey Model}
\usage{
lotkaVolterra(perturb = 1.05, viewcode = FALSE)
}
\arguments{
\item{perturb}{is a multiplier that is used to perturb the model away from equilibrium.  The default 1.05 perturbs by 5 percent away from equilibrium}

\item{viewcode}{TRUE or FALSE (default) indicating whether to print the function code}
}
\value{
a list containing the jacobian matrix and dominant eigenvalue, and creates figure 4.8
}
\description{
Simulate the time dynamics of the Lotka-Volterra model, and recreate figure 4.8
}
\examples{
# Recreate figure 4.8 exactly
lotkaVolterra(perturb = 1.05)
# create a larger initial perturbation away from equilibrim
lotkaVolterra(perturb = 1.25)
# View the code
lotkaVolterra(viewcode = TRUE)
}
