% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runge.kutta.R
\name{runge.kutta}
\alias{runge.kutta}
\title{4th Order Runge-Kutta}
\usage{
runge.kutta(Xstart, modelpars, fun, deltat, n.years)
}
\arguments{
\item{Xstart}{is a vector or array of the starting values of the ODE}

\item{modelpars}{is a list object with all model parameters}

\item{fun}{is the name of the function that calculates dXdt, where X is a vector of state variables}

\item{timesteps}{is a vector of time increments starting at t=0}
}
\value{
matrix with time in first column and state variables in remaining columns
}
\description{
This function applies the 4th order Runge-Kutta method to solve a continuous time series of differential equations.  This function is useful for demonstration, but for practical purposes you might find it better to use functions in the deSolve package
}
