% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-songsparrows.R
\docType{data}
\name{songsparrows}
\alias{songsparrows}
\title{Song Sparrow Reproductive Success}
\format{
A data frame with 12 observations, each representing the number of breeding females and the number of independent offspring produced
\describe{
  \item{Year}{Observation year.}
  \item{Females}{Number of breeding females in each year.}
  \item{IndependentOffspring}{Number of offspring surviving through egg, chick, and fledgling stage.}
  }
}
\source{
Arcese, P. and Smith, J.N.M, 1988. Effects of population density and supplemental food on reproduction in song sparrows.
Journal of Animal Ecology. 57: 119 - 136.
}
\usage{
songsparrows
}
\description{
Annual (1975 - 1986) estimates of breeding success of song sparrows (Melospiza melodia) on Mandarte Island, British Columbia, Canada.
}
\keyword{datasets}
