% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wa-wolf.R
\docType{data}
\name{wa_wolf}
\alias{wa_wolf}
\title{Population Dynamics of Wolves in Washington State, U.S.A.}
\format{
A data frame with 12 observations, each representing a year
\describe{
  \item{Year}{Calendar year.}
  \item{Abundance}{Minimum number of wolves at year's end.}
  }
}
\source{
Washington Department of Fish and Wildlife, Confederated Tribes of the Colville Reservation,
Spokane Tribe of Indians, USDA-APHIS Wildlife Services, and U.S. Fish and Wildlife Service. 2021.
Washington Gray Wolf Conservation and Management 2020 Annual Report.
Washington Department of Fish and Wildlife, Ellensburg, WA, USA
}
\usage{
wa_wolf
}
\description{
Annual estimated abundance of Gray Wolves (Canus lupis) in Washington State, U.S.A.  Numbers are the "minimum estimates"
based on direct observations from ground and from air, track surveys, and camera trap surveys. Some wolves
may be present but are not detected.
}
\keyword{datasets}
