% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whaleModelSelect.R
\name{whaleModelSelect}
\alias{whaleModelSelect}
\title{Model Selection: Population Dynamics of Killer Whales}
\usage{
whaleModelSelect()
}
\value{
a matrix containing negative log likelihood, AIC, and delta AIC for each model
}
\description{
This function generates Figure 9.3 and table 9.1 by fitting a density independent, a logistic, and a beverton-holt model to the southern resident killer whale population counts.  This illustrates the estimation of population parameters assuming "process" error, and the calculation of AIC and delta AIC
}
\examples{
#view  commands
print(whale_model_select)
# generate plot in Fig 9.3 and get table 9.1 returned
AIC.table <- whale_model_select()
print(AIC.table)
}
