% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.1X.R
\name{fig2.1X}
\alias{fig2.1X}
\title{Dynamic behavior of the logistic model}
\usage{
fig2.1X(r)
}
\arguments{
\item{r}{maximum rate of population growth}
}
\description{
Generate figures 2.11 through 2.16 by specifying the maximum rate of population growth (r), and seeing accompanying cobweb diagram and the plot of population abundance vs. time
}
\examples{
#view plotting commands
print(fig2.1X)
# generate plot with "typical" dynamics
fig2.1X(r = 0.5)
# generate plot with "complex" dynamics
fig2.1X(r = 2.1)
}
