% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.3and4.R
\name{fig2.3and4}
\alias{fig2.3and4}
\title{The structure of the Logistic Model}
\usage{
fig2.3and4()
}
\description{
Generate figures 2.3 and 2.4 showing the assumed linear relationship between per-capita growth rate and population density, and the resulting relationship between Delta N Delta t and Population size.
}
\examples{
#view plotting commands
print(fig2.3and4)
# generate plot
fig2.3and4()
}
