% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.5.R
\name{fig2.5}
\alias{fig2.5}
\title{Other Density Dependent Models}
\usage{
fig2.5(viewcode = F)
}
\arguments{
\item{viewcode}{TRUE or FALSE (default) indicating whether to print the function code}
}
\description{
Generate figure 2.5 that illustrates four types of density dependent models, with plots showing N at time t+1 vs. N at time t.
}
\examples{
# generate plot
fig2.5()
 #view code
fig2.5(viewcode = TRUE)
}
