% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.5.R
\name{fig2.5}
\alias{fig2.5}
\title{Generate figure 2.5: Other Density Dependent models}
\usage{
fig2.5()
}
\description{
Generate figure 2.5: Other Density Dependent models
}
\examples{
#view plotting commands
print(fig2.5)
# generate plot
fig2.5()
}
