% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig2.6and7.R
\name{fig2.6and7}
\alias{fig2.6and7}
\title{Compare logistic to Allee effect model}
\usage{
fig2.6and7()
}
\description{
Generate figure 2.6 and 2.7, comparing standard compensatory model (logistic) with one that has depensatory population growth / Allee Effects
}
\examples{
#view plotting commands
print(fig2.6and7)
# generate plot
fig2.6and7()
}
