% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig5.3.R
\name{fig5.3}
\alias{fig5.3}
\title{Effect of Stochasticity on Average Population Size}
\usage{
fig5.3()
}
\description{
Demonstration of how variability in population growth rate leads to reduced average population size (Figure 5.2). In each simulation, annual population growth rate, lambda, varies randomly within a specified range.  This is repeated many times, and the average over each simulation is plotted.
}
\examples{
# View the code
print(fig5.3)
# View Figure 5.3
fig5.3()
}
