% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig8.4.R
\name{fig8.4}
\alias{fig8.4}
\title{Poison Maximum Likelihood Estimation, multiple observations}
\usage{
fig8.4()
}
\value{
a list object containing the maximum likelihood estimate of r and the confidence interval
}
\description{
This function generates  Figure 8.4 and returns the maximum likelihood estimate of r and the confidence interval
}
\examples{
#view  commands
print(fig8.4)
# generate plot shown in Fig 8.4
fig8.4()
$mle
[1] 0.2
$ci
[1] 0.12 0.31

}
