% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig8.8.R
\name{fig8.8}
\alias{fig8.8}
\title{Fit Survival Function to Goby Data}
\usage{
fig8.8()
}
\value{
a list object containing the maximum likelihood estimate of pmax and alpha
}
\description{
This function generates Figure 8.8 by estimating the maximum likelihood parameter values of the survival function in Eq. 8.10
}
\examples{
#view  commands
print(fig8.8)
# generate plot in Fig 8.8 and get maximum likelihood estimated
fit <- fig8.8()
print(fit)
$pmax.mle
[1] 0.5002728
$alpha.mle
[1] 0.01271631
}
