% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicholsonBailey.R
\name{nicholsonBailey}
\alias{nicholsonBailey}
\title{Nicholson-Bailey Host Parasitoid Model}
\usage{
nicholsonBailey(perturb = 1.01)
}
\arguments{
\item{perturb}{is a multiplier that is used to perturb the model away from equilibrium.  The default 1.01 perturbs by 1 percent away from equilibrium}
}
\value{
a list object containing jacobian matrix, dominant eigenvalue, and eigenvalue magnitude
}
\description{
This function simulates the dynamics of the Nicholson Bailey Host Parasitoid model, plots Fig. 4.9, and returns the Jacobian matrix, dominant eigenvalue and eigenvalue magnitude
}
\examples{
#view  commands
print(nicholsonBailey)
# generate plot with equivalent to Fig 4.9
nicholsonBailey()
# generate a plot with larger perturbation away from equilibrium
nicholsonBailey(perturb = 1.05)
}
