% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantecol-package.R
\docType{package}
\name{quantecol-package}
\alias{quantecol}
\alias{quantecol-package}
\title{quantecol: Introduction to Quantitative Ecology}
\description{
Data and example R scripts used in "Introduction to Quantitative Ecology".
}
\details{
Example R code to accompany worked examples and figures shown in Introduction to Quantitative Ecology. This includes all of the datasets used, and many of the functions are named based on the figure number within the book text.  You may find it useful to view the code within the functions, which you can see by using the "print" function.  For instance, if you wish to view to code that was used to generate figure 2.2, you can enter print(fig2.2) into the console.
}
\keyword{internal}
